M4 release
Enable wave-center motion across several m4_vector_wave scenarios and clean up related code: toggle APPLY_MOTION -> True in annihilation*, formation*, repulsion1. Remove the hardcoded WIP launcher banner. Update formation modules to use local generate_K_positions (replace m3 import) and adjust formatting (trimmed X_NAME). Inline numpy-backed tetrahedron and K-position generator logic into m5_lagrangian_field formation10e/formation10p (add numpy import) and remove obsolete m5 formation01/02/03 parameter files. Small UI tweaks to sub-window heights and refresh methods.png image. These changes consolidate position-generation logic, enable motion behavior for experiments, and remove deprecated configs.
