+++ openwave/xperiments/m4_vector_wave/__init__.py
@@ -1,5 +1,5 @@
 """
-[WIP] M4/ VECTOR-WAVE Method (analytical tools)
+M4/ VECTOR-WAVE Method (analytical tools)
 
 Experiments exploring wave dynamics using vector-field method simulation.
 """
