+++ openwave/xperiments/m4_vector_wave/xparameters/annihilation1.py
@@ -41,7 +41,7 @@ XPARAMETERS = {
             [0.3, 0.0, 0.0],
             [-0.3, 0.0, 0.0],
         ],
-        "APPLY_MOTION": False,  # Toggle to apply motion at wave-centers, from force at each iteration
+        "APPLY_MOTION": True,  # Toggle to apply motion at wave-centers, from force at each iteration
     },
     "ui_defaults": {
         "SHOW_AXIS": False,  # Toggle to show/hide axis lines
