+++ openwave/xperiments/m4_vector_wave/xparameters/repulsion1.py
@@ -29,7 +29,7 @@ XPARAMETERS = {
         ],
         # Phase offsets for each wave-center (integer degrees, converted to radians internally)
         "PHASE_OFFSETS_DEG": [0, 0],
-        "APPLY_MOTION": False,  # Toggle to apply motion at wave-centers, from force at each iteration
+        "APPLY_MOTION": True,
     },
     "ui_defaults": {
         "SHOW_AXIS": False,  # Toggle to show/hide axis lines
